Restructure the fixtures
